+++ src/database/db.ts
@@ -1,5 +1,5 @@
 import { getDatabase } from "firebase/database";
-import app from "../services/firebase";
+import app from "../config/firebase";
 
 const database = getDatabase(app);
 export { database };
